Improve tsc and build setup - Remove vite and vitest types from globals - separate tsconfig for tests - move tests out of src - Use `tsc -b` for building
