+++ package.json
@@ -28,9 +28,8 @@
     "lint:fix:prettier": "prettier --write \"**/*.{ts,js,cjs,md}\"",
     "lint:eslint": "eslint",
     "lint:fix:eslint": "eslint --fix",
-    "build": "npm run tsc",
-    "tsc": "tsc",
-    "watch": "tsc -w",
+    "build": "tsc -b",
+    "watch": "tsc -b -w",
     "test": "vitest",
     "test:perf": "vitest run flex.performance.spec.ts",
     "prepare": "husky",
